+++ tests/commands/test_users.py
@@ -23,7 +23,7 @@ import pytest
 
 from croud.api import Client, RequestMethod
 from croud.users.commands import transform_roles_list
-from tests.util import assert_rest, call_command
+from tests.util import assert_rest, call_command, gen_uuid
 
 pytestmark = pytest.mark.usefixtures("config")
 
@@ -90,3 +90,17 @@ def test_transform_roles_list(mock_request):
         "project-2: project_member,\n"
         "project-3: project_member"
     )
+
+
+@mock.patch.object(Client, "request", return_value=(None, {}))
+def test_users_delete(mock_request, capsys):
+    user_id = gen_uuid()
+    with mock.patch("builtins.input", side_effect=["Y"]) as mock_input:
+        call_command("croud", "users", "delete", "--user-id", user_id)
+    assert_rest(mock_request, RequestMethod.DELETE, f"/api/v2/users/{user_id}/")
+    mock_input.assert_called_once_with(
+        "Are you sure you want to delete the user? [yN] "
+    )
+
+    _, err_output = capsys.readouterr()
+    assert "User deleted." in err_output