+++ CHANGES.rst
@@ -7,6 +7,8 @@ Unreleased
 
 - Dropped support for Python 3.7.
 
+- Added support for deleting users.
+
 1.5.0 - 2023/07/27
 ==================